refactor(indexeddb): rename module (indexeddb_serializer -> serializer)
