+++ crates/matrix-sdk-indexeddb/src/serializer.rs
